refactor: StaffTableクラスを追加
